Create README.md
In the field of welding quality inspection, it is always a challenge to obtain a large number and variety of marking data. In order to solve this problem, this study successfully constructed a data set containing a variety of weld defect samples, named DAR. The establishment of this data set has significantly improved the problems of insufficient number of samples and single type in previous public data sets, and provided valuable resources for the research of weld defect detection.