Update README.md
Within the realm of welding quality inspection, acquiring an ample and diverse array of labeled data has consistently posed a significant challenge. To address this issue, our research has successfully developed a comprehensive dataset teeming with varied weld defect samples, which we have designated as DAR. The introduction of this dataset has markedly alleviated the prevalent shortcomings of previous publicly available datasets, which were characterized by a dearth of samples and a lack of diversity. Consequently, the DAR dataset has become an invaluable asset in the advancement of research pertaining to the detection of weld defects.

The WIF and DAR datasets will be made public soon.